feature: Adição de esquema zod para validação de itens, e correções para corrigir as rotas de itens, com base em erros capturados pelos testes

+++ src/repositories/Item.ts
@@ -59,8 +59,15 @@ export default class ItemRepository implements ItemInterfaceRepository {
     }
     async delete(id: number): Promise<{ success: boolean; message: string }> {
         try {
-            await this.repository.delete(id)
+            const itemToDelete = await this.repository.findOneBy({id: id});
 
+            if (!itemToDelete){
+                return {
+                    "success": false,
+                    "message": "Nenhum Item encontrado com esse id!"
+                }
+            }
+            await this.repository.delete(id)
             return {
                 success: true,
                 message: "Item deletado com sucesso!"
